refactor(providers): split user-preference providers out of app_providers.dart
First batch of #4506 (app_providers.dart decomposition). Moves five
user-preference providers — feedAspectRatio, audioSharing, aiTraining,
audioDevice, language — to a new `lib/providers/preferences_providers.dart`.
`app_providers.dart` re-exports the new file so the 295 existing consumer
import sites stay unchanged.

Picked as the validation batch because the cluster is self-contained
(only depends on `sharedPreferencesProvider`), has no bridge providers,
and exercises both the `Provider<X>(...)` and `@Riverpod` codegen paths.

Refs #4506, #4339 Wave 2
